Update functionality, docs, and bug fixes
The way Next.js works--with all its optimizations--doesn't allow to have
a token on the first render, we could do it with SSR, but that would
compromise the home page. Maybe doing it in a login page would be ideal.

We can't access the cookie from the client because it's saved in
an httpOnly cookie. For this reason, we can't do Double-submit cookie
pattern since we can't get the token from the cookie to add it to the
header. Again, we could add it with `getInitialProps` but we compromise
important optimizations.

So far, we are able to do Synchronizer Token Pattern and with the caveat
of doing a `/csrf` endpoint, the thing that the Understanding CSRF
document suggest against.