Setup e2e testing
Next.js maintainers recommend to use e2e for testing endpoints, to run a
Next.js app instance and send a request against the API route directly.

* https://github.com/vercel/next.js/discussions/15166#discussioncomment-37824

For running the webserver and running the e2e I decided to use
Playwright.

Also added yalc for managing local next-csrf as a dependency locally in the
example folder.